cleanup_t

+++ include/CommonConstants.hpp
@@ -2,23 +2,24 @@
 #define COMMON_CONSTANTS_HPP_
 
 #include <cstddef>
+#include <cstdint>
 
 namespace constants {
 
-constexpr int kChunkSize = 16;
-constexpr int kChunkNumHorizontal = 200;
-constexpr int kChunkNumVertical = 100;
+constexpr int32_t kChunkSize = 16;
+constexpr int32_t kChunkNumHorizontal = 200;
+constexpr int32_t kChunkNumVertical = 100;
 
-constexpr int kWorldWidth = kChunkSize * kChunkNumHorizontal;
-constexpr int kWorldHeight = kChunkSize * kChunkNumVertical;
+constexpr int32_t kWorldWidth = kChunkSize * kChunkNumHorizontal;
+constexpr int32_t kWorldHeight = kChunkSize * kChunkNumVertical;
 
-constexpr double kWantedFramesPerSecond = 60.0;
+constexpr float kWantedFramesPerSecond = 60.0;
 constexpr size_t kFpsBufferSize = 120;
 
-constexpr int kDebugDigitPrecision = 4;
-constexpr double kDebugRealEpsilon = 1e-7;
+constexpr int32_t kDebugDigitPrecision = 4;
+constexpr float kDebugRealEpsilon = 1e-7;
 
-constexpr double kWantedSecondsPerFrame = 1.0 / kWantedFramesPerSecond;
+constexpr float kWantedSecondsPerFrame = 1.0 / kWantedFramesPerSecond;
 
 constexpr float kGravityAcceleration = 0.4F;
